categorie update

+++ src/main/java/school/token/hackaton_groupe7/application/exceptions/EntityNotFoundException.java
@@ -1,7 +1,7 @@
 package school.token.hackaton_groupe7.application.exceptions;
 
 public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(Class entityClass, Long id) {
+    public EntityNotFoundException(Class entityClass, Integer id) {
         super("Entity " + entityClass.getSimpleName() + " with id " + id + " not found");
     }
 }
