+++ src/main/java/school/token/hackaton_groupe7/application/features/categorieUser/queries/CategorieUserQueryProcessor.java
@@ -4,15 +4,14 @@ import org.springframework.stereotype.Service;
 import school.token.hackaton_groupe7.application.features.categorieUser.queries.getAll.CategorieUserGetAllOutput;
 import school.token.hackaton_groupe7.application.features.categorieUser.queries.getById.CategorieUserGetByIdOutput;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
 import school.token.hackaton_groupe7.application.utils.IQueryHandler;
 
 @Service
 public class CategorieUserQueryProcessor {
     private final IQueryHandler<Pageable, CategorieUserGetAllOutput> getAllHandler;
-    private final IQueryHandler<Long, CategorieUserGetByIdOutput> getByIdHandler;
+    private final IQueryHandler<Integer, CategorieUserGetByIdOutput> getByIdHandler;
 
-    public CategorieUserQueryProcessor(IQueryHandler<Pageable, CategorieUserGetAllOutput> getAllHandler, IQueryHandler<Long, CategorieUserGetByIdOutput> getByIdHandler) {
+    public CategorieUserQueryProcessor(IQueryHandler<Pageable, CategorieUserGetAllOutput> getAllHandler, IQueryHandler<Integer, CategorieUserGetByIdOutput> getByIdHandler) {
         this.getAllHandler = getAllHandler;
         this.getByIdHandler = getByIdHandler;
     }
@@ -22,7 +21,7 @@ public class CategorieUserQueryProcessor {
     }
 
 
-    public CategorieUserGetByIdOutput getById(Long id) {
+    public CategorieUserGetByIdOutput getById(Integer id) {
         return getByIdHandler.handle(id);
     }
 }
